requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
web3==6.11.3
hyperon==0.1.12
python-multipart==0.0.6
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.5.0
redis==4.6.0
celery==5.3.4
aioredis==2.0.1
httpx==0.25.2
ipfshttpclient==0.8.0a2
anthropic==0.25.0
bcrypt==4.1.2
setuptools==69.0.0
twilio==8.10.0
pyotp==2.9.0
